First cut implementation of huffman coded qoi files
Qoi files can now be optionally huffman coded, for a size reduction of ~14% (qoibench.c testsuit), making it basically the same as PNG files but at a speed cost of ~30%. Still way faster than libpng and stbi, but not as great as I hoped.
The qoi header stays not encoded, and the MSB colorspace byte now indicates a huffman encoded file.
The huffman encoding converts each byte into a code word of variable bit length based on occurrence frequency. A histogram of frequencies is recorded during qoi encoding and common bytes will be represented by short words. Words can have a max length of 32bits. If longer words are encountered, or the compression rate would not be sufficient (<=3%), huffman coding is skipped and a plain qoi file is generated.
A header with the words and their length is inserted after the qoi header decoding.